hopr_async_runtime::prelude

Struct AbortHandle

pub struct AbortHandle { /* private fields */ }

A handle to an Abortable task.

Implementations

impl AbortHandle

pub fn new_pair() -> (AbortHandle, AbortRegistration)

Creates an (AbortHandle, AbortRegistration) pair which can be used to abort a running future or stream.

This function is usually paired with a call to [Abortable::new].

impl AbortHandle

pub fn abort(&self)

Abort the Abortable stream/future associated with this handle.

Notifies the Abortable task associated with this handle that it should abort. Note that if the task is currently being polled on another thread, it will not immediately stop running. Instead, it will continue to run until its poll method returns.

pub fn is_aborted(&self) -> bool

Checks whether AbortHandle::abort was called on any associated AbortHandles, which includes all the AbortHandles linked with the same [AbortRegistration]. This means that it will return true even if:

abort was called after the task had completed.

abort was called while the task was being polled - the task may still be running and will not be stopped until poll returns.

This operation has a Relaxed ordering.