hopr_async_runtime

Struct AbortableList

pub struct AbortableList<T>(/* private fields */);

List of Abortable tasks with each task identified by a unique key of type T.

Abortable objects, such as join or abort handles, do not by design abort when dropped. Sometimes this behavior is not desirable, and spawned run-away tasks may still continue to live e.g.: after an error is raised.

This object allows safely managing abortable tasks and will terminate all the tasks in reverse insertion order once dropped.

Additionally, this object also implements Abortable allowing it to be arbitrarily nested.

Implementations

impl<T: Hash + Eq> AbortableList<T>

pub fn insert<A: Abortable + Send + Sync + 'static>( &mut self, process: T, task: A, )

Appends a new abortable task to the end of this list.

pub fn abort_one(&mut self, process: &T) -> bool

Looks up a task by its key, removes it and aborts it.

Returns true if the task was aborted and removed. Otherwise, returns false (including a situation when the task was present but already aborted).

pub fn extend_from(&mut self, other: AbortableList<T>)

Extends this list by appending other.

The tasks from other are moved to this list without aborting them. Afterward, other will be empty.

pub fn flat_map_extend_from<U>( &mut self, other: AbortableList<U>, key_map: impl Fn(U) -> T, )

Extends this list by appending other while mapping its keys to the ones in this list.

The tasks from other are moved to this list without aborting them. Afterward, other will be empty.

impl<T> AbortableList<T>

pub fn is_empty(&self) -> bool

Checks if the list is empty.

pub fn size(&self) -> usize

Returns the number of abortable tasks in the list.

pub fn iter_names(&self) -> impl Iterator<Item = &T>

Returns an iterator over the task names in the insertion order.

pub fn abort_all(&self)

Aborts all tasks in this list in the reverse insertion order.

Skips tasks which were already aborted.