hopr_async_runtime

Module prelude

Structs

JoinHandle

A handle that awaits the result of a task.

Functions

cancel_join_handle

sleep

Sleeps for the specified amount of time.

spawn

Spawns a task.

spawn_blocking

Spawns a blocking task.

spawn_local

Spawns a task onto the thread-local executor.

timeout_fut

Awaits a future or times out after a duration of time.